+++ src/ts-prototype/index.ts
